Add support for interprocedural tracking through summaries.
The transfer function for the nullability analysis now consults the
current Module summary to see if values passed as arguments to a call
are not-nullable.  If they are, record them as also not-nullable in
the caller.

There are only rudimentary tests so far.

It should work for mutually recursive functions (SCCs) but it isn't
tested yet.