Add code to deal with calls to external functions (through summaries).  Includes some tests.
The summary code works great.  This includes a really simple library
with a summary (base.json) that is loaded for all tests.  The transfer
function for external calls seems to work great.

There is a test case in this commit that currently
fails (unchecked-after-not-null-call.c).  The idea of the fix is
straightforward, but the mechanics might need a bit of work.
Basically, a call to a function with an argument that is nullable
should prevent that value from being marked as not-nullable later
on (since the callee does the null check).  Tracking a set of such
values should suffice.