When constructing visualization reports, use ByteString instead of Text
We don't know what encoding the source files will be in, so don't
pretend they are unicode.  This also lets us do fewer encode/decode
steps.

An alternative, if we really want to pretend the inputs are utf8,
we could use decodeUtf8With and replace invalid characters.

This won't work at all with utf16, of course