Allow allocators to call their underlying allocator through a function pointer under certain conditions.
If the function pointer is global and has a default value that is also
an allocator, consider the caller to be an allocator (assuming the
other allocator conditions are met).  This allows a common pattern
where a library allows its malloc to be overridden with a custom
external allocator.

It isn't completely safe because the user could switch in whatever
function they want, but it is a clear statement of intent by the
library author and supports the common case.