Improve the precision of the allocator and finalizer analyses
Uses an auxillary analysis, SingleInitializer, that identifies
access paths and globals that are only initialized once to a function
value in the library.

This covers the common case of libraries that support user-replacable
allocators.  We don't know what the user will pass, but we know that
it should have the semantics of the default that the library chose.