Do not check finalizer arguments past arity
If an actual argument to a call is passed at an arity greater than
that expected by the callee, do not look up its summary.  This avoids
a crash and doesn't lose any information because the callee cannot
use extra arguments for anything.

This happens if the callee is casted to a different unrelated type.