Incrementally generate ctypes output
This change generates bindings roughly a function at a time, rather
than keeping the entire Module in memory at once.  Memory usage is
consistently low, and concatenating all of the output is cheap
due to the underlying use of blaze-builder.

Making the binding for GSL used to take over a gig of ram; now it
takes only a few MB (resident).