Add a new escape annotation: PAFptrEscape
This builds on the updated escape analysis in llvm-analysis and allows
for more fine-grained treatment of escaping parameters.  This will be
useful in the code generator to provide a "safe" codegen mode that
handles all escapes equally.  It can then provide an "unsafe" mode to
make bindings more convenient by ignoring PAFptrEscapes.

A large number of escapes are of this form, so this should have a very
positive effect when used by the code generators.