Make tests runnable again.
The base.json summaries for two of the test suites needed to be
updated with libraryEnums fields.  The others needed to be switched to
use the high-level dependency loader that reads the standard library
descriptions from memory instead of on disk, to match up with recent
changes to the frontend.