Fixed the missing ref counted type problem.
Some types were being dropped in the code that matched metadata up
with the corresponding struct type while trying to find types that are
exposed in library interfaces via parameters and return values.

Now, if the step doing the matching fails, the type is just treated as
opaque instead of being dropped entirely.