Recognize a larger class of reference counted types.
Some simple types do not ever store to their finalizer field.  In
these cases, fall back to identifying reference counted types based on
structural subtyping.  Treat all subtypes of the type passed to the
unref function as reference counted.

> void unref(Foo* x){ /* ... */ }

That is, all structural subtypes of Foo are reference counted.

This changes the PAUnref annotation to record a list of relevant
structural supertypes.