Extend the outparam analysis to recognize allocations through out params.
This adds a new annotation PAOutAlloc for out parameters that are
freshly allocated by an allocator, and for which the caller assumes
ownership.

This required some extensions to the escape analysis to let the
analysis require that out parameter allocations do not escape via any
assignments besides the one creating the out parameter.