Rip out the type canonicalization hack
I pushed type canonicalization down to the IR translation layer, so
this ad-hoc version is no longer required.  The low level version is
more efficient and more precise, and other analyses benefit from it,
too.

No regressions observed in the test suite or in real-world libraries
with the change.  In fact, it even fixes some corner cases that the
ad-hoc algorithm didn't handle.