Make contract escapes work in the escape analysis.
These are escapes where a value is passed to a function pointer, but
we know some possible targets of the indirect call.  Looking at *just*
these targets, we conclude that they represent an interface (contract)
that the library expects for those function pointers.

Therefore, we say that the value does not escape *unless* the contract
is broken (and the user substitutes in a function that has different
semantics).

The test cases are derived from libarchive.