Another rewrite of the escape analysis
This version uses the same algorithm but is backed by an explicit graph
instead of datalog.  The datalog implementation was precise but too slow
(it did not complete on simple benchmarks like pcre).  This commit makes
it as fast as the original, but using the working algorithm.  It is still
considerably simpler than the original, too.