Improve alias handling in the HTML reports
Now only list functions that are actually visible - before the report
generator was only looking at function linkage and not including
visibility.  This mean that functions marked with visibility(hidden)
showed up in the public interface.

For most libraries it doesn't matter much, but for others (like
libxml2), there are public aliases to expose these functions and the
private variants are decorated with strange mangled names.  This change
makes everything look like a proper public API.