Vast simplification of the error analysis
This is a step towards implementing the four simple rules in the
module description.  Learning new functions based on return codes is
not yet implemented (that is tricky and will require a lot of work and
maybe machine learning).

The results are fairly similar to what we had before.  There is still
the strange 0 error code leaking in to our results that shouldn't
be there for libarchive.