Add a missing witness to the generalization case that is implemented
This is a rough one to say which call we generalized from (the
instruction is wrong - it is the last in the basic block instead of
the call).  This could be made more precise at the cost of a bit
of code duplication or maybe just a find.

This case also never fires right now because the errorFunctions set
is never populated