Enable compound predicate checking to identify error handling code
See the extensive comment in the new function (relevantInducedFacts).
The idea is basically to use all of the facts that we know must hold
at a given program point when doing the check for additional precision.

This change also modifies the SMT query to be an isSatisfiable check
rather than an isTheorem check, fixing some latent problems and
accommodating these compound predicates.