Modify error code learning heuristic.
We always want to generate an error descriptor when we see a known
error-reporting function called.  However, we only want to learn a new
error code if it is not a common one (like 1 or 0, perhaps others).
Before, error descriptors were not generated if one of these common
error codes was returned.