Ignore dependency function calls in error actions.
This change addresses error handling code like the following:

> set_error(errno, strlen(bar), formatMsg(".."));

We want to consider this as a single function call since the
"outermost" call set_error uses the other function call results.  That
is, they were not called for their effects but their results.

This pattern is reasonably common.  Seeing this as an action with
three function calls introduced a few problems in the analysis because
some heuristics only work with single-function call error actions.

Includes test case.