Augment the Symbolic Access Path analysis
Now, arguments written into (some path into) the return value of a
function are recorded in a new structure: the sapReturnContainer.
This is a may analysis (so a phi node that /could/ have an argument
stored in it is record).

This information is used to synthesize new paths based on arguments
when the return value of a function is stored into some path of an
argument.