Hide Escape annotations on non-pointer types.
That they show up is kind of an artifact of the analysis that could be
managed - just not showing them is a bit easier, though.  The
"incorrectly" annotated values do flow to escaping locations in the
value flow escape graph, but that isn't a useful piece of information
for scalars.