Bugfix in the rewrite of the error handling analysis
Now require candidate blocks for the known error handling case to
have only a single predecessor.  This was the intent before but I
forgot to actually check the condition.

Additionally, this change corrects the error summaries for fread and
fwrite.  We do not actually have summaries for these functions because
they do not really return error codes.  Instead, users hoping to catch
stdio errors have to call ferror.  I added a summary for that function
and corrected the expected results of the relevant tests.

Those tests were basically incorrect before.