Tweak SAT checking
The general form of queries is

   errorCondition && knownFactsAtLocation

where `errorCondition` is a formula that is true when an
error is being handled.  The idea is to see if that and the facts
in scope are consistent.  If they are, an error is being handled.
The code was not dealing with the case where no (relevant) facts are
known at the current scope.  In those cases, `errorCondition` was
trivially satisfiable, which is not really the desired meaning.

This change fails early (even before SMT) if no facts are known at the
current location.