Lift the single-predecessor restriction on known-errors
Before, the analysis would not consider blocks with more than one
predecessor.  This was for simplicity, but made us miss some inferences
we should have been able to make.  That is corrected with this change.

Note that the analysis that identifies success codes still has this
restriction.  That could, in principle, be lifted in a similar way.
It may not be necessary, however.