More refined handling of output parameters
Reading the value of a pointer makes it an input parameter.  Example:

> if(p)
>   *p = 0;

The main idea is that conditions on pointer values allow the user to
choose what code is being executed, and we don't want to automatically
choose for them.  This is especially important in the presence of
guarded side effects (imagine if there was a call on the RHS on the
assignment).

This change also introduces a flag to enable a "hack" behavior where
we allow the code above to report @p@ as an output parameter because
there are no other side effects in the guarded branch.