Move output parameter allocation to the Allocators analysis
It makes more sense here and that is how things are written up
in my dissertation.  Now the output parameter analysis is very low-level
and the allocator analysis depends on it.

One test (transitive output allocation) does not pass.  I know how to
fix it, and the next commit will take care of that.