Invert the svm dependency
Since foreign-inference uses TemplateHaskell, we can't link in
svm-simple directly (because it uses C++ and the TH linker dies).
Inverting the dependency so that the caller of 'identifyErrorHandling'
code has to provide the classifier function.  The iiglue binary
can safely link against svm-simple, so it will provide a function
instead of just providing the classifier data blob.