Restructure the analysis phases for the error-reporting analysis
Before, each analysis was run over functions one at a time.  This
order was problematic because the known-error analysis relies on
information from the known-success analysis.  It would be useful to
have all of the known-success information available before looking for
failures.

This change applies each analysis in phases (over all functions).

This change also pushes the transitive error analysis to come after
the generalization pass has run so that it can collect as much
information as possible.