Add a more selective model of success
Now, in order to be a success indicator, a function whose error
return is ignored must be followed by some other non-return
non-phi non-direct-branch code.  The idea is to try to avoid flagging
cleanup code with an ignored return code as a success.

It may be worth expanding this so that a basic block of cleanup code
(e.g. free; free; free;) could also be handled.