Aggressively remove successes from error descriptors
Once we learn something is a success (by the new majority criterion),
never consider a success code to be an error code, even if we
previously thought so.

This works because we aren't classifying boolean functions anymore.

This change also modifies how error code summaries are used.  Before,
only the first FAReportsErrors annotation would be used.  Now all are
used.

Furthermore, for transitive errors, the error action is dropped because
only the callee takes that action - the caller does not.