Improve a highlighting regex
This regular expression replaces function calls in code snippets with
links to the called function.  This tweaks the regex to make sure the
matched word ends at a word boundary and doesn't just highlight some
prefix of a word.  This showed up in various places, but an example
was deflate/deflateEnd, where deflate highlighted a prefix of
deflateEnd, preventing deflateEnd from being correctly highlighted.